ADD
Logica de negocio y corecciones.

+++ AtlantidaApplication/Models/DataBase/CLIENTEXPRODUCTO.cs
@@ -7,8 +7,8 @@
         public int idproducto { get; set; }
         public DateTime? fchsolicitud { get; set; }
         public DateTime? fchaprobacion { get; set; }
-        public float? saldoaprobado { get; set; }
-        public float? saldodisponible { get; set; }
+        public double? saldoaprobado { get; set; }
+        public double? saldodisponible { get; set; }
         public string estado { get; set; }
     }
 }
