+++ AtlantidaApplication/Models/DataBase/PRODUCTO.cs
@@ -6,8 +6,8 @@
         public string codigo { get; set; }
         public string? nombre { get; set; }
         public string? descripcion { get; set; }
-        public float? ptjinteres { get; set; }
-        public float? ptjsaldominimo { get; set; }
+        public double? ptjinteres { get; set; }
+        public double? ptjsaldominimo { get; set; }
         public int? rngminimo { get; set; }
         public int? rngmaximo { get; set; }
         public string estado { get; set; }
